+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Filter ordered content sets by profile field values.
 - Add list_button_title to Whatsapp messages.
+- Add ability to specify actions to Whatsapp list item messages.
 
 ### Removed
 - Removed word embeddings search (`s` query parameter in API)